% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDNAseq.chm13v2.R
\docType{data}
\name{chm13v2.1kbp.PE150}
\alias{chm13v2.1kbp.PE150}
\title{chm13v2.1kbp.PE150}
\format{
A annotated data frame with pre-calcuated metrics for each bin describing:
\describe{
\item{chromosome}{Chromosome name}
\item{start}{Bin start position}
\item{end}{Binend position}
\item{bases}{Percentage of non-N nucleotides acros the full bin}
\item{gc}{Percentage of C and G nucleotides \(of non-N nucleotides\)}
\item{mappability}{Average mappability of  with a maximum of 2 mismatches}
\item{blacklist}{Percent overlap with High-Signal regions from ExcludeRanges}
\item{residual}{Median loess residual from 1000 genomes \strong{PENDING}}
\item{use}{Weather bin exceeds minimum thresholds for analysis}
}
}
\source{
\url{https://github.com/RodrigoGM/QDNAseq.chm13v2}
}
\usage{
data(chm13v2.1kbp.PE150)
}
\description{
Bin annotations for kbp using  reads
}
\author{
RGM
}
\keyword{datasets}
