% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDNAseq.chm13v2.R
\docType{data}
\name{used.fastq.manifest}
\alias{used.fastq.manifest}
\title{used.fastq.manifest}
\format{
A annotated data frame with pre-calcuated metrics for each bin describing:
\describe{
\item{url}{FASTQ download URL}
\item{md5}{MD5 sums}
\item{Data.collection}{Data Collection at 1000 Genomes Project}
\item{Data.type}{type of data}
\item{Analysis.group}{PCR-free high coverage}
\item{Sample}{Sample Name}
\item{Population}{Population represented in the 1000 Genomes Project}
}
}
\source{
\url{https://github.com/RodrigoGM/QDNAseq.chm13v2}
}
\usage{
data(used.fastq.manifest)
}
\description{
Manifest of FASTQ sequences used to estimate residuals
}
\author{
RGM
}
\keyword{datasets}
